+++ CHANGELOG.md
@@ -103,7 +103,7 @@
 - fix: 1634  fix image description service [\#1667](https://github.com/elizaOS/eliza/pull/1667) ([nusk0](https://github.com/nusk0))
 - fix: Separated imageModelProvider and imageVisionModelProvider for ImageDescriptioServices [\#1664](https://github.com/elizaOS/eliza/pull/1664) ([denizekiz](https://github.com/denizekiz))
 - fix: Update Supabase schema.sql [\#1660](https://github.com/elizaOS/eliza/pull/1660) ([0xRider](https://github.com/0xRider))
-- fix: Fix bug in plugin-bootstrap/src/evaluators/facts.ts [\#1648](https://github.com/elizaOS/eliza/pull/1648) ([metakai1](https://github.com/metakai1))
+- fix: Fix bug in plugin-message-handling/src/evaluators/facts.ts [\#1648](https://github.com/elizaOS/eliza/pull/1648) ([metakai1](https://github.com/metakai1))
 - Added Hungarian README [\#1645](https://github.com/elizaOS/eliza/pull/1645) ([mdominikd](https://github.com/mdominikd))
 - fix: lockfile wasn't updated after dependency + bring viem to root level package.json [\#1642](https://github.com/elizaOS/eliza/pull/1642) ([monilpat](https://github.com/monilpat))
 - chore: update viem dependency version in plugin-evm and plugin-goat [\#1637](https://github.com/elizaOS/eliza/pull/1637) ([bertux](https://github.com/bertux))